requirements.txt
black==24.8.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pipdeptree==2.23.4
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
